+++ webapp/src/app/components/widgets/event-widget/event-registry/event-registry.service.ts
@@ -22,6 +22,7 @@ import { StartNpcTradeMenu } from './start-npc-trade-menu';
 import { Wait } from './wait';
 import { ShowSideMsg } from './show-side-msg';
 import { ShowModalChoice } from './show-modal-choice';
+import { SetMsgExpression } from './set-msg-expression';
 
 type EventConstructor<T extends EventType> = new (domSanitizer: DomSanitizer, data: T, actionStep: boolean) => AbstractEvent<T>;
 
@@ -54,6 +55,7 @@ export class EventRegistryService {
 		this.register('OPEN_QUEST_DIALOG', OpenQuestDialog);
 		this.register('SHOW_SIDE_MSG', ShowSideMsg);
 		this.register('SHOW_MODAL_CHOICE', ShowModalChoice);
+		this.register('SET_MSG_EXPRESSION', SetMsgExpression);
 		
 	}